+++ NEWS.md
@@ -1,5 +1,6 @@
 # midnight (development version) 0.1.1.90x
 
+- Added `mixup()` for creating synthetic data using Mixup algorithm.
 - Reduced dependencies to other packages.
 - Added `fastLmMatrix()`, a wrapper function of `RcppEigen::fastLm()` supporting matrix responses.
 - Renamed and Updated the `ggmid.midimp` wrapper (for better alignment with **midr** 0.6.0).